Moved resources module under ocs module
