+++ tests/test_pod_exec.py
@@ -1,7 +1,7 @@
 import os
 os.sys.path.append(os.path.dirname(os.getcwd()))
 
-from resources import pod
+from ocs.resources import pod
 
 
 def test_main():
