+++ tox.ini
@@ -14,7 +14,6 @@ commands = py.test \
     -c pytest_unittests.ini \
     --cov=ocs \
     --cov=ocsci \
-    --cov=resources \
     --cov=utility \
     {posargs}
